+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-04-18 16:55:56
-// Commit: feat(i18n): add async i18n stream provider
+// Updated at: 2026-04-21 23:06:21
+// Commit: feat(i18n): integrate offline caching for i18n state
 
 class i18nComponentState {
-  final String commitRef = '526';
-  final String lastMessage = 'feat(i18n): add async i18n stream provider';
-  final DateTime timestamp = DateTime.parse('2026-04-18T16:55:56+05:00');
+  final String commitRef = '551';
+  final String lastMessage = 'feat(i18n): integrate offline caching for i18n state';
+  final DateTime timestamp = DateTime.parse('2026-04-21T23:06:21+05:00');
 }